fix bad imports from src/core/process.ts
