+++ src/execute/jupyter/jupyter-kernel.ts
@@ -1,9 +1,8 @@
 /*
-* jupyter-kernel.ts
-*
-* Copyright (C) 2020-2022 Posit Software, PBC
-*
-*/
+ * jupyter-kernel.ts
+ *
+ * Copyright (C) 2020-2022 Posit Software, PBC
+ */
 
 import { existsSync } from "fs/mod.ts";
 import { join } from "path/mod.ts";
@@ -11,7 +10,8 @@ import { error, info, warning } from "log/mod.ts";
 
 import { sleep } from "../../core/async.ts";
 import { quartoDataDir, quartoRuntimeDir } from "../../core/appdirs.ts";
-import { execProcess, ProcessResult } from "../../core/process.ts";
+import { execProcess } from "../../core/process.ts";
+import { ProcessResult } from "../../core/process-types.ts";
 import { md5Hash } from "../../core/hash.ts";
 import { resourcePath } from "../../core/resources.ts";
 import { pythonExec } from "../../core/jupyter/exec.ts";
